@@ -31,7 +31,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
 
         <pkg.implementationTitle>ThingsBoard Performance Tests</pkg.implementationTitle>
         <pkg.mainClass>org.thingsboard.tools.PerformanceTestApplication</pkg.mainClass>
@@ -46,21 +46,21 @@
         <tb-ce-performance-test.docker.name>tb-ce-performance-test</tb-ce-performance-test.docker.name>
         <docker.push-arm-amd-image.phase>pre-integration-test</docker.push-arm-amd-image.phase>
 
-        <thingsboard.version>3.4.0</thingsboard.version>
-        <spring-boot.version>2.7.0</spring-boot.version>
-        <spring.version>5.3.20</spring.version>
-        <lombok.version>1.18.22</lombok.version>
-        <commons-lang3.version>3.4</commons-lang3.version>
-        <netty.version>4.1.75.Final</netty.version>
-        <netty-tcnative-boringssl-static.version>2.0.51.Final</netty-tcnative-boringssl-static.version>
-        <bouncycastle.version>1.68</bouncycastle.version>
+        <thingsboard.version>4.0.0</thingsboard.version>
+        <spring-boot.version>3.1.5</spring-boot.version>
+        <spring.version>6.0.13</spring.version>
+        <lombok.version>1.18.30</lombok.version>
+        <commons-lang3.version>3.12.0</commons-lang3.version>
+        <netty.version>4.1.100.Final</netty.version>
+        <netty-tcnative-boringssl-static.version>2.0.61.Final</netty-tcnative-boringssl-static.version>
+        <bouncycastle.version>1.76</bouncycastle.version>
         <winsw.version>2.0.1</winsw.version>
-        <slf4j.version>1.7.32</slf4j.version>
-        <logback.version>1.2.10</logback.version>
+        <slf4j.version>2.0.9</slf4j.version>
+        <logback.version>1.4.11</logback.version>
         <californium.version>2.6.3</californium.version>
         <leshan.version>2.0.0-M4</leshan.version>
         <dockerfile-maven.version>1.4.13</dockerfile-maven.version>
-        <awaitility.version>4.1.0</awaitility.version>
+        <awaitility.version>4.2.0</awaitility.version>
     </properties>
 
     <build>
@@ -87,10 +87,10 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.8.1</version>
+                    <version>3.11.0</version>
                     <configuration>
-                        <source>11</source>
-                        <target>11</target>
+                        <source>17</source>
+                        <target>17</target>
                     </configuration>
                 </plugin>
                 <plugin>
